requirements.txt
pandas
scikit-learn
joblib
matplotlib
seaborn
pyyaml
numpy